Before:

```
sequential inserts: 79.037ms
inserts: 141.911ms
range queries: 783.089ms
splices: 6839.141ms
deletes: 49.744ms
```

After:

```
sequential inserts: 72.125ms
inserts: 83.611ms
range queries: 593.052ms
splices: 3616.571ms
deletes: 72.975ms
```

Everything is faster except for deletions, which are still quite fast. I was surprised to see that sequential insertions were not improved as significantly as random insertions.

/cc @nathansobo