@nathansobo 2016-12-30
Looks good. I started playing with something similar in gaps of time during the holiday but I like this implementation. I'm still interested in the following questions:

* Would a small vector optimization be helpful where we store the values directly in the handle struct for sets of small cardinality? It seems like sets will be extremely small, on the order of 1 or 2 elements, for a large number of use cases, and this would avoid indirections and allocations. Not sure how big a win it would be though.
* Would treating the sets as *values* and using `std::set_union` / `std::set_difference` on entire sets in the rotations be helpful? That would replace lots of random insertions with a sequence of append operations. Again, not sure if this would really matter until the sets got fairly large, which should be uncommon.